+++ src/ehrdata/__init__.py
@@ -6,7 +6,3 @@ from .core import EHRData
 __all__ = ["EHRData", "dt", "io", "pl", "pp", "tl"]
 
 __version__ = version("ehrdata")
-
-from .logging_config import configure_logging
-
-configure_logging()
